[refactor][m]: implement efficient gql management system (#41)
