+++ components/resource/DataExplorer.tsx
@@ -1,30 +1,9 @@
-import Link from 'next/link';
-import ErrorMessage from '../Error';
 import { useQuery } from '@apollo/react-hooks';
-import gql from 'graphql-tag';
-
-const QUERY = gql`
-  query dataset($id: String) {
-    dataset(id: $id) @rest(type: "Response", path: "package_show?{args}") {
-      result {
-        resources {
-          name
-          id
-          title
-          description
-          format
-          size
-          created
-          last_modified
-          url
-        }
-      }
-    }
-  }
-`;
+import ErrorMessage from '../Error';
+import { GET_RESOURCES_QUERY } from '../../graphql/queries';
 
 export default function DataExplorer({ variables }) {
-  const { loading, error, data } = useQuery(QUERY, {
+  const { loading, error, data } = useQuery(GET_RESOURCES_QUERY, {
     variables,
     // Setting this value to true will make the component rerender when
     // the "networkStatus" changes, so we are able to know if it is fetching
