+++ components/search/List.tsx
@@ -1,28 +1,10 @@
+import { useQuery } from '@apollo/react-hooks';
 import Item from './Item';
 import ErrorMessage from '../Error';
-import { useQuery } from '@apollo/react-hooks';
-import gql from 'graphql-tag';
-
-const QUERY = gql`
-  query search($q: String, $sort: String) {
-    search(q: $q, sort: $sort)
-      @rest(type: "Search", path: "package_search?{args}") {
-      result {
-        results {
-          name
-          title
-          organization {
-            name
-            title
-          }
-        }
-      }
-    }
-  }
-`;
+import { SEARCH_QUERY } from '../../graphql/queries';
 
 export default function List({ variables }) {
-  const { loading, error, data } = useQuery(QUERY, {
+  const { loading, error, data } = useQuery(SEARCH_QUERY, {
     variables,
     // Setting this value to true will make the component rerender when
     // the "networkStatus" changes, so we are able to know if it is fetching
