+++ components/static/List.tsx
@@ -1,19 +1,10 @@
-import ErrorMessage from '../Error';
 import parse from 'html-react-parser';
 import { useQuery } from '@apollo/react-hooks';
-import gql from 'graphql-tag';
-
-const QUERY = gql`
-  query posts {
-    posts @rest(type: "Posts", path: "", endpoint: "wordpress-posts") {
-      found
-      posts
-    }
-  }
-`;
+import ErrorMessage from '../Error';
+import { GET_POSTS_QUERY } from '../../graphql/queries';
 
 export default function List() {
-  const { loading, error, data } = useQuery(QUERY, {
+  const { loading, error, data } = useQuery(GET_POSTS_QUERY, {
     // Setting this value to true will make the component rerender when
     // the "networkStatus" changes, so we are able to know if it is fetching
     // more data
